Improve calendar and form
