+++ components/Calendar.tsx
@@ -1,14 +1,15 @@
 import useTranslation from 'next-translate/useTranslation'
 import calcPrice from '../helpers/calcPrice'
+import getWeeks from '../helpers/getWeeks'
 import styles from './Calendar.module.scss'
 
 const defaultSubs = {
-  petita: {},
-  mitjana: {},
-  gran: {},
-  ous: {},
   ceba: {},
   fruita: {},
+  gran: {},
+  mitjana: {},
+  ous: {},
+  petita: {},
 }
 
 function Calendar({
@@ -19,26 +20,34 @@ function Calendar({
 }) {
   const { t, lang } = useTranslation('my-baskets')
   const weeks = getWeeks(lang)
+  const isActive = (s, f, i) => s[f].count > 0 && i % parseInt(s[f].time) === 0
 
   return (
     <div className={styles.calendar} {...props}>
       {weeks.map((week, index) => {
-        const sub = exceptions[week.name] || subscription
-        const pActive = index % parseInt(sub.petita.time) === 0
-        const mActive = index % parseInt(sub.mitjana.time) === 0
-        const gActive = index % parseInt(sub.gran.time) === 0
-        const oActive = index % parseInt(sub.ous.time) === 0
-        const fActive = index % parseInt(sub.fruita.time) === 0
-        const cActive = index % parseInt(sub.ceba.time) === 0
+        const sub = exceptions[week.id] || subscription
+        const cActive = isActive(sub, 'ceba', index)
+        const fActive = isActive(sub, 'fruita', index)
+        const gActive = isActive(sub, 'gran', index)
+        const mActive = isActive(sub, 'mitjana', index)
+        const oActive = isActive(sub, 'ous', index)
+        const pActive = isActive(sub, 'petita', index)
         const active = pActive || mActive || gActive
 
         return (
           <div
-            key={week.name}
-            title={active ? 'Editar contingut setmana' : undefined}
-            onClick={
-              active ? () => onClickSubscription({ ...sub, week }) : undefined
-            }
+            key={week.id}
+            title={active ? t`edit` : undefined}
+            onClick={() => {
+              const daySub = { ...sub }
+              Object.keys(daySub).forEach((k) => {
+                daySub[k] = {
+                  count: isActive(daySub, k, index) ? daySub[k].count : 0,
+                  time: 1,
+                }
+              })
+              onClickSubscription({ ...daySub, week })
+            }}
             className={`${styles.day} ${active ? styles.active : ''}`}
           >
             {active ? (
@@ -95,54 +104,6 @@ function Calendar({
       })}
     </div>
   )
-}
-
-function getMonday(d) {
-  const date = new Date(d)
-  date.setHours(12)
-  const day = date.getDay()
-  const diff = date.getDate() - day + (day == 0 ? -6 : 1)
-  return new Date(date.setDate(diff))
-}
-
-function getSunday(d) {
-  const date = new Date(d)
-  date.setHours(12)
-  const today = date.getDate()
-  const dayOfTheWeek = date.getDay()
-  const newDate = date.setDate(today - dayOfTheWeek + 7)
-  return new Date(newDate)
-}
-
-function getWeek(date, lang) {
-  const locale = `${lang}-${lang.toUpperCase()}`
-
-  const monday = getMonday(date)
-  const mondayName = monday.toLocaleDateString(locale, {
-    month: 'short',
-    day: 'numeric',
-  })
-
-  const sunday = getSunday(date)
-  const sundayName = sunday.toLocaleDateString(locale, {
-    year: 'numeric',
-    month: 'short',
-    day: 'numeric',
-  })
-
-  const name = `${mondayName} - ${sundayName}`.replace(/de /g, '')
-
-  return { monday, sunday, name }
-}
-
-function getWeeks(lang = 'ca') {
-  const dayCursor = new Date()
-
-  return Array.from({ length: 12 }).map(() => {
-    const res = getWeek(dayCursor, lang)
-    dayCursor.setDate(dayCursor.getDate() + 7)
-    return res
-  })
 }
 
 export default Calendar
