+++ locales/ca/my-baskets.json
@@ -1,5 +1,8 @@
 {
   "add-button": "Afageix a la subscripció",
+  "baskets": "Cistelles",
+  "calendar": "Calendari",
+  "cancel": "Res, ja estava bé",
   "description-basket-gran": "Cistella gran de verdura i fruita ecològica de temporada. Els productes i el contingut de la cistella poden variar segons la setmana.",
   "description-basket-mitjana": "Cistella mitjana de verdura i fruita ecològica de temporada. Els productes i el contingut de la cistella poden variar segons la setmana.",
   "description-basket-petita": "Cistella pequeña de verdura i fruita ecològica de temporada. Els productes i el contingut de la cistella poden variar segons la setmana.",
@@ -10,19 +13,23 @@
   "details-extra-fruita": "Fruita ({{kg}}kg, 3 productes, {{price}}€)",
   "details-extra-ous": "Ous ({{price}}€ la mitja dotzena)",
   "details-extra-patata-ceba": "Extra 2kg Patata + 1kg Ceba seca ({{price}}€)",
+  "edit": "Editar contingut setmana",
+  "extras": "Extres",
   "freq-1": "Cada setmana",
   "freq-2": "Cada 2 setmanes",
   "freq-3": "Cada 3 setmanes",
-  "freq-4": "Cada mes",
+  "freq-4": "Cada 4 setmanes",
+  "next-weeks-baskets": "Cistelles programades en les pròximes setmanes",
+  "no-want": "No ho vull rebre",
+  "product-ceba": "Patata i Ceba",
+  "product-fruita": "Fruita",
   "product-gran": "Cistella gran",
   "product-mitjana": "Cistella mitjana",
-  "product-petita": "Cistella petita",
-  "product-fruita": "Fruita",
   "product-ous": "Ous",
-  "product-ceba": "Patata i Ceba",
-  "next-weeks-baskets": "Cistelles programades en les pròximes setmanes",
+  "product-petita": "Cistella petita",
   "resume": "{{kg}}kg, {{count}} productes, {{price}}€",
+  "save": "Desar canvis",
   "title": "Les meves cistelles",
-  "view-calendar": "Veure calendari",
+  "when": "Cada quan ho vol rebre?",
   "what-include": "Aquesta setmana la cistella inclou:"
 }
