+++ locales/es/my-baskets.json
@@ -1,5 +1,8 @@
 {
   "add-button": "Añadir a la subscripción",
+  "baskets": "Cestas",
+  "calendar": "Calendario",
+  "cancel": "Nada, ya estaba bien",
   "description-basket-gran": "Cesta grande de verdura y fruta ecológica de temporada. Los productos y el contenido de la cesta pueden variar según la semana.",
   "description-basket-mitjana": "Cesta mediana de verdura y fruta ecológica de temporada. Los productos y el contenido de la cesta pueden variar según la semana.",
   "description-basket-petita": "Cesta pequeña de verdura y fruta ecológica de temporada. Los productos y el contenido de la cesta pueden variar según la semana.",
@@ -10,19 +13,23 @@
   "details-extra-fruita": "Fruta ({{kg}}kg, 3 productos, {{price}}€)",
   "details-extra-ous": "Huevos ({{price}}€ la media docena)",
   "details-extra-patata-ceba": "Extra 2kg Patata + 1kg Cebolla ({{price}}€)",
+  "edit": "Editar contenido semana",
+  "extras": "Extras",
   "freq-1": "Cada semana",
   "freq-2": "Cada 2 semanas",
   "freq-3": "Cada 3 semanas",
-  "freq-4": "Cada mes",
+  "freq-4": "Cada 4 semanas",
+  "next-weeks-baskets": "Cestas programadas para las próximas semanas",
+  "no-want": "No lo quiero recibir",
+  "product-ceba": "Patata y Cebolla",
+  "product-fruita": "Fruta",
   "product-gran": "Cesta grande",
   "product-mitjana": "Cesta mediana",
-  "product-petita": "Cesta pequeña",
-  "product-fruita": "Fruta",
   "product-ous": "Huevos",
-  "product-ceba": "Patata y Cebolla",
-  "next-weeks-baskets": "Cestas programadas para las próximas semanas",
+  "product-petita": "Cesta pequeña",
   "resume": "{{kg}}kg, {{count}} productos, {{price}}€",
+  "save": "Guardar cambios",
   "title": "Mis cestas",
-  "view-calendar": "Ver calendario",
+  "when": "Cada cuando lo quieres recibir?",
   "what-include": "Esta semana la cesta incluye:"
 }
